refactor: extract branches visualization to AlertBranches component
- Create AlertBranches.svelte following AlertRequestLogs pattern
- Add alertBranches() trigger function to alert.ts
- Update SideChatList to use new trigger function
- Remove 70+ lines of inline code from AlertComp.svelte

Note: Branch visualization feature had pre-existing issues before this refactor
